+++ boards/apis/llm/gemini/v1beta3/tests/generativelanguage.models.list.js
@@ -1,6 +1,4 @@
 import { base, board } from "@google-labs/breadboard";
-
-import { starter } from "@google-labs/llm-starter";
 import { core } from "@google-labs/core-kit";
 
 /*
@@ -26,6 +24,6 @@ export default await board(() => {
     path: "./spec/generativelanguage.models.embedText.json",
     input: input.api_inputs,
     ...input,
-    ...starter.secrets({ keys: ["GEMINI_API_KEY"] }),
+    ...core.secrets({ keys: ["GEMINI_API_KEY"] }),
   });
 }).serialize(metaData);