+++ packages/breadboard-cli/src/commands/lib/utils.ts
@@ -6,7 +6,7 @@
 
 import { BoardRunner, GraphDescriptor } from "@google-labs/breadboard";
 import { Dirent, watch as fsWatch } from "fs";
-import { opendir, readFile, stat, writeFile } from "fs/promises";
+import { opendir, readFile, stat, writeFile, mkdir } from "fs/promises";
 import { join } from "node:path";
 import { stdin as input } from "node:process";
 import * as readline from "node:readline/promises";
@@ -72,17 +72,18 @@ export const loadBoard = async (
   file: string,
   options: Options
 ): Promise<BoardRunner> => {
-  const loaderType = extname(file).slice(1) as "js" | "ts" | "yaml" | "json";
+  const loaderType = extname(file).slice(1) as "js" | "ts" | "json";
   const save = "save" in options ? options["save"] : true;
 
   const loader = new Loaders(loaderType);
   const board = await loader.load(file, options);
-  if (save && loaderType !== "json") {
-    const pathInfo = path.parse(file);
+  if (save) {
     const boardClone = JSON.parse(JSON.stringify(board));
-    const outputFilePath = path.join(options.output, `${pathInfo.name}.json`);
     delete boardClone.url; // Boards shouldn't have URLs serialized.
     const boardJson = JSON.stringify(boardClone, null, 2);
+
+    // Most commands will pass in the output directory, but if they don't, we'll use the directory of the file being loaded.
+    const outputFilePath = await resolveAndCreateOutputDirectory(options, file);
     await writeFile(outputFilePath, boardJson);
   }
   return board;
@@ -98,6 +99,11 @@ export const parseStdin = async (): Promise<string> => {
   return lines;
 };
 
+const showError = (e: unknown, path?: string) => {
+  const error = e as Error;
+  console.error(`Failed to load board at "${path}": ${error.message}`);
+};
+
 export const loadBoards = async (
   path: string,
   options: Options
@@ -106,37 +112,45 @@ export const loadBoards = async (
   const fileUrl = pathToFileURL(path);
 
   if (fileStat && fileStat.isFile() && path.endsWith(".json")) {
-    const data = await readFile(path, { encoding: "utf-8" });
-    const board = JSON.parse(data) as GraphDescriptor; // assume conversion would fail if it wasn't a graph descriptor.
+    try {
+      const data = await readFile(path, { encoding: "utf-8" });
+      const board = JSON.parse(data) as GraphDescriptor; // assume conversion would fail if it wasn't a graph descriptor.
 
-    return [
-      {
-        edges: board.edges ?? [],
-        nodes: board.nodes ?? [],
-        kits: board.kits ?? [],
-        title: board.title ?? path,
-        url: join("/", relative(process.cwd(), path)),
-        version: board.version ?? "0.0.1",
-      },
-    ];
+      return [
+        {
+          edges: board.edges ?? [],
+          nodes: board.nodes ?? [],
+          kits: board.kits ?? [],
+          title: board.title ?? path,
+          url: join("/", relative(process.cwd(), path)),
+          version: board.version ?? "0.0.1",
+        },
+      ];
+    } catch (e) {
+      showError(e, path);
+    }
   }
 
   if (
     fileStat &&
     fileStat.isFile() &&
-    (path.endsWith(".js") || path.endsWith(".ts") || path.endsWith(".yaml"))
+    (path.endsWith(".js") || path.endsWith(".ts") || path.endsWith(".json"))
   ) {
-    // Compile the JS, TS or YAML.
-    const board = await loadBoard(path, options);
+    try {
+      // Compile the JS or TS.
+      const board = await loadBoard(path, options);
 
-    return [
-      {
-        ...board,
-        title: board.title ?? path,
-        url: join("/", relative(process.cwd(), path)),
-        version: board.version ?? "0.0.1",
-      },
-    ];
+      return [
+        {
+          ...board,
+          title: board.title ?? path,
+          url: join("/", relative(process.cwd(), path)),
+          version: board.version ?? "0.0.1",
+        },
+      ];
+    } catch (e) {
+      showError(e, path);
+    }
   }
 
   if (fileStat && fileStat.isDirectory()) {
@@ -153,6 +167,22 @@ const getFilename = (dirent: Dirent) => {
   return maybePath.endsWith(name) ? maybePath : join(maybePath, name);
 };
 
+async function resolveAndCreateOutputDirectory(options: Options, file: string) {
+  const outputRoot =
+    "output" in options ? options["output"] : path.dirname(file);
+  const inputBaseDirectory = options.root
+    ? path.resolve(options.root)
+    : path.dirname(file);
+
+  const pathInfo = path.parse(file);
+  const fullOutputRootPath = path.resolve(outputRoot);
+  const inputRoot = path.relative(inputBaseDirectory, pathInfo.dir);
+  const base = path.join(fullOutputRootPath, inputRoot);
+  const outputFilePath = path.join(base, `${pathInfo.name}.json`);
+  await mkdir(base, { recursive: true });
+  return outputFilePath;
+}
+
 async function loadBoardsFromDirectory(
   fileUrl: URL,
   path: string,
@@ -161,38 +191,32 @@ async function loadBoardsFromDirectory(
   const dir = await opendir(fileUrl);
   const boards: Array<BoardMetaData> = [];
   for await (const dirent of dir) {
-    if (dirent.isFile() && dirent.name.endsWith(".json")) {
-      const data = await readFile(getFilename(dirent), {
-        encoding: "utf-8",
-      });
-      const board = JSON.parse(data);
-      boards.push({
-        ...board,
-        title: board.title ?? join("/", getFilename(dirent)),
-        url: join("/", getFilename(dirent)),
-        version: board.version ?? "0.0.1",
-      });
-    }
-
     if (
       dirent.isFile() &&
       (dirent.name.endsWith(".js") ||
         dirent.name.endsWith(".ts") ||
-        dirent.name.endsWith(".yaml"))
+        dirent.name.endsWith(".json"))
     ) {
-      const path = getFilename(dirent);
-      const board = await loadBoard(path, options);
-      boards.push({
-        ...board,
-        title: board.title ?? join("/", path),
-        url: join("/", path),
-        version: board.version ?? "0.0.1",
-      });
+      const filename = getFilename(dirent);
+      try {
+        const board = await loadBoard(filename, options);
+        boards.push({
+          ...board,
+          title: board.title ?? join("/", filename),
+          url: join("/", filename),
+          version: board.version ?? "0.0.1",
+        });
+      } catch (e) {
+        showError(e, filename);
+      }
     }
 
     if (dirent.isDirectory()) {
+      const baseFolder = fileUrl.pathname.endsWith("/")
+        ? fileUrl.pathname
+        : `${fileUrl.pathname}/`;
       const boardsInDir = await loadBoardsFromDirectory(
-        new URL(dirent.name, fileUrl),
+        new URL(dirent.name, pathToFileURL(baseFolder)),
         join(path, dirent.name),
         options
       );