+++ packages/breadboard-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [9bcd607]
+- Updated dependencies [f6a7f43]
+  - @google-labs/breadboard@0.10.0
+
 ## 0.1.5
 
 ### Patch Changes