+++ packages/website/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @google-labs/breadboard-website
 
+## 2.0.1
+
+### Patch Changes
+
+- 0c106e0: Adds TOC, header links, and callouts
+
 ## 2.0.0
 
 ### Major Changes